\name{addPosition}
\alias{addPosition}
\title{Add position data to parsed data}
\arguments{
  \item{data}{data.frame of parsed data}
}
\value{
  annotated data.frame
}
\description{
  Adds and x and a y column to parsed data.  These are used
  for laying out tiles in a tile plot.  Values are
  organized by lane, then by swath and surface.
}

