\name{buildReports}
\alias{buildReports}
\title{Generate Illumina reports folder}
\arguments{
  \item{project}{SAV project}

  \item{destination}{relative location to save reports
  folder}
}
\description{
  Generate a folder of images that approximates the format
  of the folder that was superceded by InterOp.
}

