% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/savR-methods.R
\docType{methods}
\name{clusterQualityGtN}
\alias{clusterQualityGtN}
\alias{clusterQualityGtN,savProject,integer,integer,integer}
\alias{clusterQualityGtN,savProject,integer,integer,integer-method}
\title{Get the proportion of clusters over a specified quality threshold}
\usage{
clusterQualityGtN(project, lane, cycle, n)

\S4method{clusterQualityGtN}{savProject,integer,integer,integer}(project, lane,
  cycle, n = 30L)
}
\arguments{
\item{project}{SAV project}

\item{lane}{lane(s) number}

\item{cycle}{cycle(s) number}

\item{n}{quality threshold}
}
\description{
Return the ratio of clusters with a quality score less than or equal to
a specified value (n) for the requested lanes and cycles.
}
\examples{
\dontrun{
example(savR)
clusterQualityGtN(fc, 1L, 25L, 30L)
}
}

