% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/SavR-accessors.R
\docType{methods}
\name{clusters}
\alias{clusters}
\alias{clusters,savProject,integer-method}
\title{Get number of clusters per lane}
\usage{
clusters(project, lane)

\S4method{clusters}{savProject,integer}(project, lane = 1L)
}
\arguments{
\item{project}{SAV project}

\item{lane}{lane(s) number}
}
\description{
Sum the total number of clusters for all tiles in the lane.
}
\examples{
\dontrun{
example(savR)
clusters(fc, 1L)
} 
}

