% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/SavR-accessors.R
\docType{methods}
\name{errorMetrics}
\alias{errorMetrics}
\alias{errorMetrics,savProject-method}
\title{Get Error Metrics}
\usage{
errorMetrics(project)

\S4method{errorMetrics}{savProject}(project)
}
\arguments{
\item{project}{SAV project}
}
\value{
sorted data.frame of Error metrics
}
\description{
Error metrics for lane, tile, and cycle.
}
\details{
\describe{
 \item{\code{lane}:}{Lane number}
 \item{\code{tile}:}{Tile ID}
 \item{\code{cycle}:}{Cycle number}
 \item{\code{errorrate}:}{Error rate}
 \item{\code{nPerfect}:}{number of perfect reads}
 \item{\code{n[1-4]Error}:}{Number of reads with 1, 2, 3 and 4 errors}
}
}
\examples{
example(savR)
colnames(extractionMetrics(fc))
}

