\name{illuminaFlowCellLayout}
\alias{illuminaFlowCellLayout}
\title{Layout of an Illumina flowcell}
\description{
  Layout of an Illumina flowcell
}
\section{Slots}{
  \describe{ \item{\code{lanecount}:}{Number of lanes on
  the flowcell} \item{\code{surfacecount}:}{Number of
  surfaces} \item{\code{swathcount}:}{Number of imaging
  swaths} \item{\code{tilecount}:}{Number of tiles per
  swath} }
}

