\name{illuminaRead}
\alias{illuminaRead}
\title{Illumina read}
\description{
  Illumina read
}
\section{Slots}{
  \describe{ \item{\code{number}:}{the index of this read
  in sequencing} \item{\code{cycles}:}{number of cycles in
  this read} \item{\code{index}:}{logical representing
  whether or not this read is an index read} }
}

