% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/savR-methods.R
\docType{methods}
\name{indexFrequencies}
\alias{indexFrequencies}
\alias{indexFrequencies,savProject-method}
\title{Get Index Frequency table}
\usage{
indexFrequencies(project)

\S4method{indexFrequencies}{savProject}(project)
}
\arguments{
\item{project}{SAV project}
}
\description{
Returns a data frame of index sequences and percentage of clusters passing the filter.
}
\details{
\describe{
 \item{\code{sample}:}{The sample ID assigned to an index in the sample sheet.}
 \item{\code{project}:}{The project assigned to an index in the sample sheet.}
 \item{\code{I7}:}{The sequence for the first index read (i7).}
 \item{\code{I5}:}{The sequence for the second index read (i5).}
 \item{\code{cluster}:}{Number of clusters passing filter that are assigned to the index.}
 \item{\code{perc}:}{Percentage of clusters passing filter that are assigned to the index.}
}
}
\examples{
\dontrun{
example(savR)
indexFrequencies(fc)
}
}
