% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/SavR-accessors.R
\docType{methods}
\name{indexMetrics}
\alias{indexMetrics}
\alias{indexMetrics,savProject-method}
\title{Get Index Metrics}
\usage{
indexMetrics(project)

\S4method{indexMetrics}{savProject}(project)
}
\arguments{
\item{project}{SAV project}
}
\value{
data.frame of index metrics sorted according to order in file (same as in IPA-6541)
}
\description{
Lane, tile, read, index, cluster, sample, project
}
\details{
\describe{
 \item{\code{lane}:}{Lane number}
 \item{\code{tile}:}{Tile ID}
 \item{\code{read}:}{Read number}
 \item{\code{index}:}{Index name}
 \item{\code{cluster}:}{Index cluster count}
 \item{\code{sample}:}{Sample name}
 \item{\code{project}:}{Project name}
}
}
\examples{
example(savR)
colnames(indexMetrics(fc))
}
