% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/savR-methods.R
\docType{methods}
\name{plotIndexFrequencies}
\alias{plotIndexFrequencies}
\alias{plotIndexFrequencies,savProject,logical-method}
\title{Plot Index Frequencies}
\usage{
plotIndexFrequencies(project, dataLabels)

\S4method{plotIndexFrequencies}{savProject,logical}(project,
  dataLabels = FALSE)
}
\arguments{
\item{project}{SAV project}

\item{dataLabels}{whether or not to label bars with data values, default FALSE}
}
\description{
Generate a barplot of index frequencies based on percentage of clusters passing the filter.
}
