\name{plotIntensity}
\alias{plotIntensity}
\title{Plot flowcell intensity by base and cycle}
\arguments{
  \item{project}{A \link{savProject} object}

  \item{cycle}{integer cycle number}

  \item{base}{character for nucleotide}
}
\description{
  Draws a representation of a flowcell, showing the scaled
  corrected intensity values.
}

