% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sav.R
\name{read_sav_header}
\alias{read_sav_header}
\title{Reads headers and sample ID list from SAV file.}
\usage{
read_sav_header(path)
}
\arguments{
\item{path}{SAV file path.}
}
\value{
A list of headers and sample IDs.
}
\description{
Reads headers and sample ID list from SAV file.
}
