% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sav.R
\name{read_sav_region}
\alias{read_sav_region}
\title{Reads the specified region from a SAV file.}
\usage{
read_sav_region(path, chrom, beg = 0, end = 2147483647, fmt = "GT")
}
\arguments{
\item{path}{SAV file path.}

\item{chrom}{Chromosome to query.}

\item{beg}{Start position (Default: 0).}

\item{end}{End position (Default: 2147483647).}

\item{fmt}{Whether to read data as genotypes, allele counts, haplotype dosages or dosages (GT, AC, HDS, DS, Default: GT).}
}
\value{
A data frame of site info and a matrix of genotype data.
}
\description{
Reads the specified region from a SAV file.
}
