% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/SavR-accessors.R
\docType{methods}
\name{reads}
\alias{reads}
\alias{reads,savProject-method}
\title{Get reads}
\usage{
reads(project)

\S4method{reads}{savProject}(project)
}
\arguments{
\item{project}{SAV project}
}
\value{
List of \link{illuminaRead-class} objects
}
\description{
Accessor to obtain information about the reads of a particular Illumina
sequencing run.
}
\examples{
example(savR)
reads(fc)
}

