\docType{class}
\name{savCorrectedIntensityFormat-class}
\alias{savCorrectedIntensityFormat-class}
\title{Corrected Intensity formatter}
\description{
Lane, tile, cycle, average intensity, corrected intensities
(ACGT), average corrected called intensities (ACGT), number
of no-calls, number of (ACGT) calls, and signal to noise
ratio.
}
\section{Slots}{
  \describe{ \item{\code{name}:}{vector of column names}
  \item{\code{type}:}{vector of data types of elements}
  \item{\code{lengths}:}{vector of byte lengths for each
  element} \item{\code{order}:}{vector of column names for
  sorting} \item{\code{version}:}{integer version number} }
}

