\name{savCorrectedIntensityFormat}
\alias{savCorrectedIntensityFormat}
\title{Corrected Intensity formatter}
\description{
  Lane, tile, cycle, average intensity, corrected
  intensities (ACGT), average corrected called intensities
  (ACGT), number of no-calls, number of (ACGT) calls, and
  signal to noise ratio.
}

