% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{savData-class}
\alias{savData-class}
\title{Structure for holding parsed InterOp headers and data}
\description{
Structure for holding parsed InterOp headers and data
}
\section{Slots}{

\describe{
\item{\code{header}:}{list of parsed header values}
\item{\code{data}:}{data.frame of parsed values}
}
}

