% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{savErrorFormat-class}
\alias{savErrorFormat-class}
\title{Error Metrics formatter}
\description{
Lane, tile, cycle, errorrate, nPerfect, n1Error, n2Error,
n3Error, n4Error.
}
\section{Slots}{

\describe{
\item{\code{name}:}{vector of column names}
\item{\code{type}:}{vector of data types of elements}
\item{\code{lengths}:}{vector of byte lengths for each element}
\item{\code{order}:}{vector of column names for sorting}
\item{\code{version}:}{integer version number}
}
}

