% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{savExtractionFormat-class}
\alias{savExtractionFormat-class}
\title{Extraction Metrics formatter}
\description{
Lane, tile, cycle, FWHM (ACGT), intensity (ACGT), datestamp, timestamp.
Datestamp and timestamp are munged at the moment because R does not
have native support for 32-bit unsigned integers and I have not implemented
a solution.
}
\section{Slots}{

\describe{
\item{\code{name}:}{vector of column names}
\item{\code{type}:}{vector of data types of elements}
\item{\code{lengths}:}{vector of byte lengths for each element}
\item{\code{order}:}{vector of column names for sorting}
\item{\code{version}:}{integer version number}
}
}

