\name{savExtractionFormat}
\alias{savExtractionFormat}
\title{Extraction Metrics formatter}
\description{
  Lane, tile, cycle, FWHM (ACGT), intensity (ACGT),
  datestamp, timestamp. Datestamp and timestamp are munged
  at the moment because R does not have native support for
  32-bit unsigned integers and I have not implemented a
  solution.
}

