% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{savIndexFormat-class}
\alias{savIndexFormat-class}
\title{Index Metrics formatter}
\description{
Lane, tile, read, index, cluster, sample, project
}
\section{Slots}{

\describe{
\item{\code{name}:}{vector of column names}
\item{\code{type}:}{vector of data types of elements}
\item{\code{lengths}:}{vector of byte lengths for each element}
\item{\code{order}:}{vector of column names for sorting}
\item{\code{version}:}{integer version number}
}
}
