\docType{class}
\name{savQualityFormat-class}
\alias{savQualityFormat-class}
\title{Quality Metrics formatter}
\description{
Lane, tile, cycle, Q1-Q50 counts
}
\section{Slots}{
  \describe{ \item{\code{name}:}{vector of column names}
  \item{\code{type}:}{vector of data types of elements}
  \item{\code{lengths}:}{vector of byte lengths for each
  element} \item{\code{order}:}{vector of column names for
  sorting} \item{\code{version}:}{integer version number} }
}

