% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{savQualityFormatV5-class}
\alias{savQualityFormatV5-class}
\title{Quality Metrics formatter version 5}
\description{
Lane, tile, cycle, Q1-Q50 counts
}
\section{Slots}{

\describe{
\item{\code{name}:}{vector of column names}
\item{\code{type}:}{vector of data types of elements}
\item{\code{lengths}:}{vector of byte lengths for each element}
\item{\code{order}:}{vector of column names for sorting}
\item{\code{version}:}{integer version number}
}
}

