\docType{package}
\name{savR-package}
\alias{savR-package}
\title{Parse and analyze Illumina SAV files}
\description{
  Parse and analyze Illumina SAV files
}
\details{
  \tabular{ll}{ Package: \tab savR \cr Type: \tab Package
  \cr Version: \tab 0.1.7 \cr Date: \tab 2013-11-18 \cr
  License: \tab AGPL-3 \cr LazyLoad: \tab yes \cr }

  Parse Illumina Sequence Analysis Viewer files, provide
  visualizations and access to data.
}
\author{
  R. Brent Calder \email{brent.calder@einstein.yu.edu}
}
\references{
  \url{http://search.cpan.org/dist/Bio-IlluminaSAV/Bio/IlluminaSAV.pm}
  \url{https://bitbucket.org/invitae/illuminate}
}
\keyword{package}

