\name{savTileFormat}
\alias{savTileFormat}
\title{Tile Metrics formatter}
\description{
  Lane, tile, code, value.  Codes are:
}
\details{
  \tabular{ll}{ 100 \tab Cluster Density \cr 101 \tab PF
  Cluster Density \cr 102 \tab Number of clusters \cr 103
  \tab Number of PF clusters \cr 400 \tab Control lane \cr
  }
}

