% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{stat_sav_file}
\alias{stat_sav_file}
\title{Get statistics about SAV file.}
\usage{
stat_sav_file(file_path)
}
\arguments{
\item{path}{SAV file path.}
}
\value{
A data frame of statistics about file.
}
\description{
Get statistics about SAV file.
}
