% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{illuminaFlowCellLayout-class}
\alias{illuminaFlowCellLayout-class}
\title{Layout of an Illumina flowcell}
\description{
Class representation of the features of an Illumina flow cell.
}
\section{Slots}{

\describe{
\item{\code{lanecount}:}{Number of lanes on the flowcell}
\item{\code{surfacecount}:}{Number of surfaces}
\item{\code{swathcount}:}{Number of imaging swaths}
\item{\code{tilecount}:}{Number of tiles per swath}
\item{\code{sectionperlane}:}{Number of sections per lane (NextSeq)}
\item{\code{lanepersection}:}{Number of lanes per section (NextSeq)}
\item{\code{tilenamingconvention}:}{Description of deviation from original formatting layout}
}
}

