% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{illuminaRead-class}
\alias{illuminaRead-class}
\title{Illumina read}
\description{
Class representation of the features of an Illumina sequencing read.
}
\section{Slots}{

\describe{
\item{\code{number}:}{the index of this read in sequencing}
\item{\code{cycles}:}{number of cycles in this read}
\item{\code{index}:}{logical representing whether or not this read is an index read}
}
}

