% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{pfClusters}
\alias{pfClusters}
\alias{pfClusters,savProject,integer-method}
\title{Get number of PF clusters per lane}
\usage{
pfClusters(project, lane)

\S4method{pfClusters}{savProject,integer}(project, lane = 1L)
}
\arguments{
\item{project}{SAV project}

\item{lane}{lane(s) number}
}
\description{
Sum the total pass filter number of clusters for all tiles in the lane.
}
\examples{
\dontrun{
example(savR)
pfClusters(fc, 1L)
}
}

