\name{plotFWHM}
\alias{plotFWHM}
\title{Generate FWHM plots}
\arguments{
  \item{project}{SAV project}

  \item{cycle}{sequence cycle}

  \item{base}{nucleotide base (ACGT)}
}
\description{
  Plots the average full width of clusters at half maximum
  (FWHM) of each tile for a given cycle and base.
}

