% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/savR-methods.R
\docType{methods}
\name{qualityHeatmap}
\alias{qualityHeatmap}
\alias{qualityHeatmap,savProject,integer,integer,logical-method}
\alias{qualityHeatmap,savProject,numeric,numeric,missing-method}
\title{Generate a heatmap of qualities}
\usage{
qualityHeatmap(project, lane, read, collapse)

\S4method{qualityHeatmap}{savProject,integer,integer,logical}(project, lane,
  read, collapse = T)

\S4method{qualityHeatmap}{savProject,numeric,numeric,missing}(project, lane,
  read)
}
\arguments{
\item{project}{SAV project}

\item{lane}{integer lane specification}

\item{read}{integer vector of sequence reads to include (not including index reads)}

\item{collapse}{whether or not to collapse index reads into the preceeding read (# reads = directions), default TRUE}
}
\description{
Plots a heatmap of quality vs cycle for a given lane for 1 or more sequence reads.  Read qualities include sequence + index.
}

