% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/savR-package.R
\docType{package}
\name{savR-package}
\alias{savR-package}
\title{Parse and analyze Illumina SAV files}
\description{
Parse Illumina Sequence Analysis Viewer files
}
\details{
\tabular{ll}{
Package: \tab savR \cr
Type: \tab Package \cr
Version: \tab 1.7.5 \cr
Date: \tab 2015-07-28 \cr
License: \tab AGPL-3 \cr
LazyLoad: \tab yes \cr
}

Parse Illumina Sequence Analysis Viewer (SAV)
files, access data, and generate QC plots.
}
\author{
R. Brent Calder \email{brent.calder@einstein.yu.edu}
}
\references{
For information about Illumina SAV, please refer to \cr \url{http://supportres.illumina.com/documents/documentation/software_documentation/sav/sequencinganalysisviewer_userguide_15020619c.pdf} \cr For other implementations (and inspiration) please see \cr \url{http://search.cpan.org/dist/Bio-IlluminaSAV/Bio/IlluminaSAV.pm} \cr \url{https://bitbucket.org/invitae/illuminate}
}
\keyword{package}

